refactor: ajuste rotas home e dashboard
